requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
mcp>=1.0.0
qdrant-client>=1.9.0
pydantic>=2.7.0
python-dotenv>=1.0.0
